Initializing A Scene With Two Player Movement
Using Godot to experiment with some mechanics.

This first scene has a default 2D sprite. It can take input from the "left" and the "right" players - the left using WASD and the right using the arrow keys. Eventually I'd like this to be done the left and right joysticks on a controller, but that's all the way upstairs right now, and I wanted to keep this fairly basic because I'm still new to Godot.

I'm primarily working on movement at the moment before visuals (and there's nothing for movement to... do, right now, anyway). Currently it feels very stiff and clunky. When players press in opposite directions, it completely stops all movement. It does have some shared functionality and teamwork in ways that I like (when two players go in the same direction, the sprite moves faster!).

Initial plan going forward will be to modify movement to something more satisfying.