Big Movement Update
Made some tweaks I was thinking about to player/players movement.

Before, a disagreement would cause a total halt in movement - no one would go anywhere. It also just felt very stiff and predictable. Now, there's a random variable (generated between -1 and 10, unsure if those will be final numbers? 10 seems a bit high) that gets multiplied by the left or right movement vector. This means that the right or left player suddenly has much more power behind their thrust (ahem) when picking a direction. It also adds a bit of chaos in terms of speed - makes it funny to be suddenly zooming off unexpectedly fast. It /feels/ fun!

Learned how to set up a timer in Godot that controls when these random multipliers shift (currently every 7 seconds). Pretty easy, it's a node and then call a function in the script to update random variables when timer runs out. I love Godot.

I like this movement mechanic in conjunction with consent and ideation of sexuality... Some times you really are pulled along by another person's desires. Often this is a way you get to discover something new that you like, and it's not always strictly consensual (in the binary yes/no sense, which... is not a particularly helpful way to think about consent, imo).