Adding more test sensors
WELL.

Wanted to build circuit and code for some new sensors to test out. Felt most intrigued by the temperature sensor and it's now the one I hate the most. Look at all that horrible useless code...

Temp sensor - because it's an analog sensor that uses heavy math there's a lot of set up for it just to use it properly. Some crazy equation I only sort of understand. Using it is... difficult. I tried many, many different ways of reading data to try and get the sort of change I wanted, but with little luck. The temperature changes immediately, which is cool - I was doing things like breathing on it to change the temperature which felt intimate and sexy - but then it takes FOREVER to reset back to a normal baseline, which feels extremely unresponsive and annoying for a controller of all things. I tried doing things like only change when a sudden change happens, taking the measurement of difference, etc, but nothing felt satisfying.

Photoresistor - used one of these before, easy, cool. Might need to tweak settings based on brightness of room. Covering with hand could be an action. Easy. More than an on/off situation since it's also analog - it takes direct readings - so maybe something possible there?

Reed sensor - i bent one of the pins to put into the breadboard and it shattered. I maintain I did nothing wrong lol. U can't use a sensor that's a straight line.

Motion sensor - idk it only had three pins, and it worked... REALLY easy? It's also an on/off switch, only on when it detects movement. It looks freaky and sci-fi. no idea if it's practical for this project.

i'm tired... i really hate that temp sensor... we fought for so long...