Tweaked sensor communicaction
Alright... Had the idea falling asleep last night that I could add a boolean flag to the temp sensor in the same way I did for the tilt sensor... That makes it more useable but I still kind of hate it.

Also messed with the photo sensor a bit so it's adaptable for different lighting in different rooms, as when I worked on it in my living room was different than TAG lab and not working as ideally. Now it should change as soon as there's a significant darkening happening (ie, someone moving a body part in front of the resistor).

Idk if having most but not all of the LED activity in a separate function is useful. This is maybe bad/confusing but for some reason I like it???