Cleaned up code, added wire touch element in code
removed unused sensor code. these sensors ended up being not useful for the project, either because the hardware didn't fit (capacitive touch sensor, motion sensor) or because hardware and code were difficult to manage (thermistor).

ledOn() got removed and LED triggers are directly in their respective functions which is more logically organized.

There's a new touch() function for making two specific needles touch together to complete a circuit. Very simple but cool. In order to avoid short circuiting the pin, voltage only runs through the pin when the needles are NOT touching, and stops when it detects a change - so a simple reversal of the logic (ie, when the pin reads False/0, it's now a "success") allows it to be useful in this controller.

All active pins are on the specific side of the board that is most easily reachable from the human controller setup.